+++ wxc_sdk/events/__init__.py
@@ -1,10 +1,10 @@
 from collections.abc import Generator
 from datetime import datetime
-from enum import Enum
 from typing import Optional, Any
 
-from wxc_sdk.api_child import ApiChild
-from wxc_sdk.base import ApiModel, dt_iso_str
+from ..api_child import ApiChild
+from ..base import ApiModel, dt_iso_str
+from ..base import SafeEnum as Enum
 
 __all__ = ['EventData', 'ComplianceEvent', 'EventResource', 'EventType', 'EventsApi']