+++ wxc_sdk/webhook/__init__.py
@@ -4,13 +4,13 @@ Webhook types and API
 import datetime
 import json
 from collections.abc import Generator
-from enum import Enum
 from typing import Optional
 
 from pydantic import Field
 
 from ..api_child import ApiChild
 from ..base import ApiModel, to_camel, webex_id_to_uuid
+from ..base import SafeEnum as Enum
 
 __all__ = ['WebHookEvent', 'WebHookResource', 'WebHookCreate', 'WebHookStatus', 'WebHook', 'WebhookApi']